% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMRgrad_generic.R
\name{UMRgrad_generic}
\alias{UMRgrad_generic}
\alias{grad_SIR_generic}
\title{Gradient of least-squares Shuffled Isotonic Regression criterion}
\usage{
UMRgrad_generic(
  yy,
  ww_y = rep(1/length(yy), length(yy)),
  mm,
  ww_m = rep(1/length(mm), length(mm)),
  AAfunc,
  BBfunc
)

grad_SIR_generic(
  yy,
  mm,
  counts = rep(1, length(mm)),
  AAfunc,
  BBfunc,
  rescale = FALSE
)
}
\arguments{
\item{yy}{Y (response) observation vector (numeric)}

\item{ww_y}{Weight vector for yy.}

\item{mm}{Current (unsorted) estimate/iterate at which to compute
gradient. (Length equals length of yy).}

\item{ww_m}{Weight vector for mm.}

\item{AAfunc}{This is the function "A" defined in the gradient
calculations in the paper (Balabdaoui, Doss, Durot (2020+).}

\item{BBfunc}{This is the function "B" defined in the gradient
    calculations in the paper (Balabdaoui, Doss, Durot (2020+).

 @details  Returns gradient as a  column matrix.  See calculations  in the
    paper.
    

 @examples
#### See help for gradDesc_PC, gradDesc, or grad_helpers}

\item{counts}{If the function that mm represents is piecewise constant, then mm may be passed in as only the unique entries.  In that case counts contains the number of times each element of mm is repeated.  Thus length(counts)==length(mm).  (Default for counts is thus a vector of all 1's.)}

\item{rescale}{Boolean: if False then the final return value is the}
}
\description{
Gradient of least-squares Shuffled Isotonic Regression criterion
}
